Security_Compliance_workshop-Deepak-Srivastava/Day1/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
boto3==1.28.57
python-dotenv==1.0.0
redis==4.6.0
gunicorn==21.2.0
